refactor: split user session routes
